Now the builds pass - broken out the fixtures
